% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nctools_spatial.R
\name{checkLongitude}
\alias{checkLongitude}
\title{Check and correct Longitude values given the position of the Prime Meridian (International Reference Meridian)}
\usage{
checkLongitude(x, primeMeridian = "center", sort = FALSE, ...)
}
\arguments{
\item{x}{numeric vector of longitude values.}

\item{primeMeridian}{position of the Prime Meridian. Use "center" for [-180,180]
range values and "left" for [0,360] range (Pacific centered).}

\item{sort}{sort longitude values?}

\item{...}{}
}
\value{

}
\description{
Check and correct Longitude values given the position of the Prime Meridian (International Reference Meridian)
}
