% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nctools_spatial.R
\name{findPrimeMeridian}
\alias{findPrimeMeridian}
\title{Find the Prime Meridian}
\usage{
findPrimeMeridian(x)
}
\arguments{
\item{x}{numeric vector of longitude values.}
}
\value{

}
\description{
Find the Prime Meridian
}
