% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nctools-extras.R
\name{nc_changePrimeMeridian}
\alias{nc_changePrimeMeridian}
\title{Change the Prime Meridian of a variable in a ncdf file.}
\usage{
nc_changePrimeMeridian(filename, output, varid = NA,
  primeMeridian = "center", verbose = FALSE, overwrite = FALSE,
  compression = NA)
}
\arguments{
\item{filename}{The filename of the original ncdf file.}

\item{output}{The name of the output file.}

\item{varid}{The name of the variable to extract.}

\item{primeMeridian}{position of the Prime Meridian. Use "center" for [-180,180]
range values and "left" for [0,360] range (Pacific centered).}

\item{verbose}{If TRUE, run verbosely.}

\item{overwrite}{overwrite output file if already exists?}
}
\description{
Change the Prime Meridian of a variable in a ncdf file.
}
