% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nctools-main.R
\name{nc_extract}
\alias{nc_extract}
\title{Create a new file with one variable from a ncdf file}
\usage{
nc_extract(filename, varid, output)
}
\arguments{
\item{filename}{The filename of the original ncdf file.}

\item{varid}{The name of the variable to extract.}

\item{output}{The name of the output file.}
}
\value{
NULL, it creates the new file in the disk.
}
\description{
Create a new file with one variable from a ncdf file
}

