% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nctools-extras.R
\name{nc_mask}
\alias{nc_mask}
\title{Extract a mask from a ncdf file.}
\usage{
nc_mask(filename, output = NULL)
}
\arguments{
\item{filename}{The filename of the original ncdf file.}

\item{output}{The file to write the mask. If NULL, the default, a list with the mask is returned.}
}
\value{
A ncdf file with the mask if output is not NULL, and a list with the mask information.
}
\description{
Extract a mask from a ncdf file.
}
