% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nctools-main.R
\name{nc_rcat}
\alias{nc_rcat}
\title{Concatenate records of the same variable from different ncdf files}
\usage{
nc_rcat(filenames, varid, output)
}
\arguments{
\item{filenames}{A vector with the file names.}

\item{varid}{The name of the variable to concatenate.}

\item{output}{The name of the output file.}
}
\value{
NULL, it creates the new file in the disk.
}
\description{
Concatenate records of the same variable from different ncdf files
}

