% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nctools-regrid.R
\name{nc_regrid}
\alias{nc_regrid}
\title{Regrid a variable from a ncdf file}
\usage{
nc_regrid(filename, varid, dim, new, mask = NULL, output, fill = FALSE,
  radius = 1, log = TRUE, ...)
}
\arguments{
\item{filename}{The filename of the original ncdf file.}

\item{varid}{The name of the variable to extract.}

\item{dim}{A vector including the names of the dimensions to interpolate,
e.g. c("longitude", "latitude")}

\item{new}{A list including information of the new coordinates, the names
should match the ones provides in \code{dim}.}

\item{mask}{An optional matrix providing the ocean/land mask. The maks will
be multiplied by each layer of the new regrided array, so dimensions must be
consistent. The mask contains 1s in the area for valid data and NAs for the
invalid ones (e.g. 1s in the ocean, NAs in the land).}

\item{output}{The filename to save the regrided variable.}
}
\description{
Regrid a variable from a ncdf file
}
