% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nctools-regrid.R
\name{nc_regrid2}
\alias{nc_regrid2}
\title{Title}
\usage{
nc_regrid2(filename, varid = NULL, dim, new, mask = NULL, output,
  fill = FALSE, radius = 1, log = TRUE, ...)
}
\arguments{
\item{...}{}
}
\description{
Title
}
