% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nctools-main.R
\name{nc_rename}
\alias{nc_rename}
\title{Renaming variable and dimensions in a netCDF File}
\usage{
nc_rename(filename, oldnames, newnames, output, verbose = FALSE)
}
\arguments{
\item{filename}{The filename of the original ncdf file.}

\item{oldnames}{A string vector containing the names of the
variable or dimensions in the file that are to be renamed.}

\item{newnames}{A string vector containing the new names of
the variables or dimensions.}

\item{output}{Optional, the output file with the changes. By default,
it will overwrite the old file.}

\item{verbose}{If TRUE, run verbosely.}
}
\description{
Renaming variable and dimensions in a netCDF File
}
