% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nctools-main.R
\name{nc_unlim}
\alias{nc_unlim}
\title{Make a dimension unlimited}
\usage{
nc_unlim(filename, unlim, output = NULL)
}
\arguments{
\item{unlim}{Name of the dimension to set as unlimited}

\item{output}{Name of the output file. If NULL,
replace the original value}
}
\value{

}
\description{
Make a dimension unlimited
}
