% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nctools-auxiliar.R
\name{ncatt_get_all}
\alias{ncatt_get_all}
\title{Get all variable's attributes from a ncdf object}
\usage{
ncatt_get_all(nc, type = c("var", "dim"))
}
\arguments{
\item{nc}{An open connection to a netCDF file as in nc_open(file).}

\item{type}{A character to choose attributes for variables ("var")
or dimensions ("dim").}
}
\value{
A list with all the atributes.
}
\description{
Get all variable's attributes from a ncdf object
}

