% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nctools-auxiliar.R
\name{ncatt_put_all}
\alias{ncatt_put_all}
\title{Put several attributes to a ncdf object}
\usage{
ncatt_put_all(nc, varid, attname, attval, prec = NA, verbose = FALSE,
  definemode = FALSE)
}
\arguments{
\item{nc}{An open connection to a netCDF file as in nc_open(file).}

\item{varid}{The variable whose attribute is to be written. Can be a
character string with the variable's name, an object of class ncvar4,
or an id contained in the "id" field of a ncvar object. As a special case,
if varid==0, then a global attribute is written instead of a variable's
attribute.}

\item{attname}{The names of the attributes.}

\item{attval}{The values of the attributes.}

\item{prec}{Precision to write the attribute.}

\item{verbose}{Can be set to TRUE if additional information is desired while
the attribute is being created.}

\item{definemode}{See 'definemode' in ncatt_put for details.}
}
\description{
Put several attributes to a ncdf object
}
