% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nctools-auxiliar.R
\name{ncdim_isUnlim}
\alias{ncdim_isUnlim}
\title{Is the dimention unlimited?}
\usage{
ncdim_isUnlim(nc)
}
\arguments{
\item{nc}{An open connection to a netCDF file as in nc_open(file).}
}
\description{
Is the dimention unlimited?
}

