% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nctools-main.R
\name{ncdim_rename}
\alias{ncdim_rename}
\title{Rename an Existing Dimension in a netCDF File}
\usage{
ncdim_rename(filename, old_dimname, new_dimname, output, verbose = FALSE)
}
\arguments{
\item{old_dimname}{A string vector containing the names of the
dimensions in the file that are to be renamed.}

\item{new_dimname}{A string vector containing the new names of
the dimensions.}

\item{output}{Optional, the output file with the changes. By default,
it will overwrite the old file.}

\item{verbose}{If TRUE, run verbosely.}
}
\description{
Rename an Existing Dimension in a netCDF File
}
