% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nctools-auxiliar.R
\name{ncdim_size}
\alias{ncdim_size}
\title{Get the dimensions of the variables in a ncdf object.}
\usage{
ncdim_size(nc)
}
\arguments{
\item{nc}{An open connection to a netCDF file as in nc_open(file).}
}
\value{
A named list with the dimensions of the variables.
}
\description{
Get the dimensions of the variables in a ncdf object.
}

