% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nctools-main.R
\name{ncdim_unlim}
\alias{ncdim_unlim}
\title{Make a dimension unlimited}
\usage{
ncdim_unlim(filename, unlim, output = NULL)
}
\arguments{
\item{unlim}{Name of the variable to set as unlimited}

\item{output}{Name of the output file. If NULL,
replace the original value}
}
\description{
Make a dimension unlimited
}

