% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nctools-auxiliar.R
\name{ncvar_change_compression}
\alias{ncvar_change_compression}
\title{Change the compression of the variables in a ncdf object.}
\usage{
ncvar_change_compression(nc, varid = NA, compression)
}
\arguments{
\item{nc}{An open connection to a netCDF file as in nc_open(file).}

\item{varid}{Either the name of the variable or an ncvar object indicating whose compression value will be changed.
If NA, all the variables will be changed to the new compression.}

\item{compression}{The new compression value.}
}
\value{

}
\description{
Change the compression of the variables in a ncdf object.
}
