% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nctools-auxiliar.R
\name{ncvar_compression}
\alias{ncvar_compression}
\title{Get the compression of the variables in a ncdf object.}
\usage{
ncvar_compression(nc, varid = NA)
}
\arguments{
\item{nc}{An open connection to a netCDF file as in nc_open(file).}

\item{varid}{Either the name of the variable or an ncvar object.}
}
\description{
Get the compression of the variables in a ncdf object.
}
