% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nctools-auxiliar.R
\name{ncvar_dim}
\alias{ncvar_dim}
\title{Get dimension names for each variable in a ncdf file}
\usage{
ncvar_dim(nc, varid = NULL, value = FALSE)
}
\arguments{
\item{nc}{An open connection to a netCDF file as in nc_open(file).}

\item{varid}{What variable to read the data from. Can be a string with the
name of the variable or an object of class ncvar4. If set to NA,
the function will determine if there is only one variable in the file and,
if so, read from that, but if there are multiple variables in the file, an error is generated.}

\item{value}{If TRUE, return a list with the value of the dimensions, not only the names of them.}
}
\description{
Get dimension names for each variable in a ncdf file
}
