% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nctools-auxiliar.R
\name{ncvar_size}
\alias{ncvar_size}
\title{Get the size of the variables in a ncdf object.}
\usage{
ncvar_size(nc, varid = NULL)
}
\arguments{
\item{nc}{An open connection to a netCDF file as in nc_open(file).}

\item{varid}{}
}
\value{

}
\description{
Get the size of the variables in a ncdf object.
}
