% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nctools-main.R
\name{write.ncdf}
\alias{write.ncdf}
\title{Data output in ncdf format}
\usage{
write.ncdf(x, filename, varid, dim, longname, units, prec = "float",
  missval = -9999, compression = 9, chunksizes = NA, verbose = FALSE,
  dim.units, dim.longname, unlim = FALSE)
}
\arguments{
\item{x}{An array to write to a ncdf file.}

\item{filename}{The file to write.}

\item{varid}{The name of the variable in the ncdf file.}

\item{dim}{A list with the values of the dimensions. Names are taken from the list.}

\item{longname}{The longname for the variable to be created.}

\item{units}{The units for the variable to be created.}

\item{prec}{The precision for the variable to be created.}

\item{unlim}{}
}
\description{
Data output in ncdf format
}
