% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nctools-main.R
\name{write_ncdf}
\alias{write_ncdf}
\title{Data output in ncdf format}
\usage{
write_ncdf(x, filename, ...)
}
\arguments{
\item{x}{An array to write to a ncdf file.}

\item{filename}{The file to write.}

\item{varid}{The name of the variable in the ncdf file.}

\item{dim}{A list with the values of the dimensions. Names are taken from the list.}

\item{longname}{The longname for the variable to be created.}

\item{units}{The units for the variable to be created.}

\item{prec}{The precision for the variable to be created.}

\item{missval}{Value set for NAs.}

\item{compression}{If set to an integer between 1 (least compression) and 9 (most compression), this enables compression for the variable as it is written to the file. Turning compression on forces the created file to be in netcdf version 4 format, which will not be compatible with older software that only reads netcdf version 3 files.}

\item{chunksizes}{For compression, the size of the chunks.}

\item{verbose}{Do you want to know what's happening?}

\item{dim.units}{Units of the dimensions}

\item{dim.longname}{Longname of the dimensions.}

\item{unlim}{Name of the unlimited dimension.}

\item{global}{List of global attributes to be saved.}
}
\value{

}
\description{
Data output in ncdf format
}
