% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nctools-regrid.R
\name{nc_regrid}
\alias{nc_regrid}
\title{Regrid a variable from a ncdf file}
\usage{
nc_regrid(
  filename,
  varid = NA,
  dim = 1:2,
  new,
  mask = NULL,
  output,
  extrap = FALSE,
  log = TRUE,
  keepRange = FALSE,
  ...
)
}
\arguments{
\item{filename}{The filename of the original ncdf file.}

\item{varid}{What variable to read the data from. Can be a string with the
name of the variable or an object of class ncvar4. If set to NA,
the function will determine if there is only one variable in the file and,
if so, read from that, but if there are multiple variables in the file, an error is generated.}

\item{dim}{A vector including the names of the dimensions to interpolate,
e.g. c("longitude", "latitude")}

\item{new}{A list including information of the new coordinates, the names
should match the ones provides in \code{dim}.}

\item{mask}{An optional matrix providing the ocean/land mask. The maks will
be multiplied by each layer of the new regrided array, so dimensions must be
consistent. The mask contains 1s in the area for valid data and NAs for the
invalid ones (e.g. 1s in the ocean, NAs in the land).}

\item{output}{The filename to save the regrided variable.}
}
\description{
Regrid a variable from a ncdf file
}
