% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nctools-main.R
\name{nc_subset}
\alias{nc_subset}
\title{Subset a ncdf variable}
\usage{
nc_subset(
  filename,
  varid,
  output,
  newvarid,
  compression,
  force_v4 = FALSE,
  ...,
  ignore.case = FALSE
)
}
\arguments{
\item{filename}{The name of the ncdf file to subset.}

\item{varid}{The name of the variable to subset. If missing and only one variable in the file, that one is used.}

\item{output}{The name of the ncdf output file to create.}

\item{newvarid}{New name for varid in the output file.}

\item{compression}{Compression level for the new variable (forces ncdf v4).}

\item{force_v4}{Logical. Should the resulting file be ncdf v4?}

\item{...}{the dimensions and bounds of values to subset.}

\item{ignore.case}{Logical. Ignore case when matching the dimensions?}
}
\value{

}
\description{
Subset a ncdf variable
}
