% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nctools-auxiliar.R
\name{ncvar_dim}
\alias{ncvar_dim}
\title{Get dimension names for each variable in a ncdf file}
\usage{
ncvar_dim(nc, varid = NULL, value = FALSE)
}
\arguments{
\item{nc}{An open connection to a netCDF file as in nc_open(file).}

\item{value}{}
}
\description{
Get dimension names for each variable in a ncdf file
}

