% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nctools-main.R
\name{write.ncdf}
\alias{write.ncdf}
\title{Data output in ncdf format}
\usage{
write.ncdf(x, filename, varid, dim, longname, units, prec = "float",
  missval = -9999, compression = 9, chunksizes = NA, verbose = FALSE,
  dim.units, dim.longname, unlim = FALSE)
}
\arguments{
\item{unlim}{}
}
\description{
Data output in ncdf format
}

